FAB-13791 Build and CI changes
This patch creates a basic Makefile to run the unit-tests
and creates Jenkinsfile with basic configuration.

Jenkinsfile run on x platform for every verify and merge changes.
It executes unit-tests on clean environment after fetch the patchset
or clone the repository. If merge job fails, it sends build notififcation
to jenkins-robot RocketChat channel.

Change-Id: I263ba738e442ec1c1ed5ac19129604fc6121db03
Signed-off-by: rameshthoomu <rameshbabu.thoomu@gmail.com>