FAB-13791 fabric-lib-go CI changes
This patch creates a Jenkinsfile to clone the refspec on the patchset
submission and clone the repository on merge event. Jenkinsfile uses the
global shared library from the ci-management repository and execute unit-tests
and send the merge build failure notifications to jenkins-robot rocketChat
channel.

Jenkinsfile run on x platform for every verify and merge changes.
It executes unit-tests on clean environment after fetch the patchset
or clone the repository. If merge job fails, it sends build notififcation
to jenkins-robot RocketChat channel.

Change-Id: I263ba738e442ec1c1ed5ac19129604fc6121db03
Signed-off-by: rameshthoomu <rameshbabu.thoomu@gmail.com>